+++ spool/settings.py
@@ -53,6 +53,7 @@ MIDDLEWARE = [
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "tracker.middleware.ForceCredentialChangeMiddleware",
     "tracker.middleware.ProfileTimezoneMiddleware",
+    "tracker.middleware.LastSeenMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]